refactor: overall refactoring
1. Improve folder structure
2. Break large component into smaller ones
3. Use arrow functions for functions expect react component
