+++ src/components/mainContent/default/trackCardList/scroll-button.tsx
@@ -1,0 +1,25 @@
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+
+import { Button } from '@/components/ui/button';
+
+function ScrollButton({
+  direction,
+  onClick,
+  disabled,
+}: {
+  direction: 'left' | 'right';
+  onClick: () => void;
+  disabled: boolean;
+}) {
+  return (
+    <Button size="icon" onClick={onClick} disabled={disabled}>
+      {direction === 'left' ? (
+        <ChevronLeft className="h-4 w-4" />
+      ) : (
+        <ChevronRight className="h-4 w-4" />
+      )}
+    </Button>
+  );
+}
+
+export default ScrollButton;
